feat: more ui stuff
