+++ Models/ConnectionStates.cs
@@ -1,0 +1,11 @@
+﻿namespace NightreignRelicEditor.Models;
+
+public enum ConnectionStates
+{
+    NotConnected,
+    NightreignNotFound,
+    EACDetected,
+    ConnectedOffsetsNotFound,
+    Connected,
+    ConnectionLost,
+}
