+++ NightreignRelicEditor.csproj
@@ -20,4 +20,21 @@
     <EmbeddedResource Include="reliceffects.tsv" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Page Update="Views\Controls\RelicData.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <XamlRuntime>Wpf</XamlRuntime>
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Update="Views\MainWindow.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <XamlRuntime>Wpf</XamlRuntime>
+      <SubType>Designer</SubType>
+    </Page>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Resources\" />
+  </ItemGroup>
+
 </Project>
