--- RelicData.xaml
@@ -1,51 +1,0 @@
-﻿<UserControl x:Class="NightreignRelicEditor.RelicData"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:local="clr-namespace:NightreignRelicEditor"
-             mc:Ignorable="d" d:DesignWidth="200">
-    <Grid Margin="10" HorizontalAlignment="Stretch">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="Auto"/>   <!-- CheckBox / Buttons -->
-            <ColumnDefinition Width="*"/>      <!-- Text -->
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>     <!-- Relic header -->
-            <RowDefinition Height="Auto"/>     <!-- Slot 1 -->
-            <RowDefinition Height="Auto"/>     <!-- Slot 2 -->
-            <RowDefinition Height="Auto"/>     <!-- Slot 3 -->
-        </Grid.RowDefinitions>
-
-        <!-- Header -->
-        <CheckBox x:Name="checkRelic1Active"
-                  Grid.Row="0" Grid.ColumnSpan="2"
-                  Content="Relic 1"
-                  IsChecked="True"
-                  FontWeight="Bold"/>
-
-        <!-- Slot 1 -->
-        <Button x:Name="buttonClearSlot1"
-                Grid.Row="1" Grid.Column="0"
-                Content="x" Width="20" Margin="0,0,5,0"/>
-        <TextBlock x:Name="textSlot1"
-                   Grid.Row="1" Grid.Column="1"
-                   Text="-" TextWrapping="Wrap"/>
-
-        <!-- Slot 2 -->
-        <Button x:Name="buttonClearSlot2"
-                Grid.Row="2" Grid.Column="0"
-                Content="x" Width="20" Margin="0,0,5,0"/>
-        <TextBlock x:Name="textSlot2"
-                   Grid.Row="2" Grid.Column="1"
-                   Text="-" TextWrapping="Wrap"/>
-
-        <!-- Slot 3 -->
-        <Button x:Name="buttonClearSlot3"
-                Grid.Row="3" Grid.Column="0"
-                Content="x" Width="20" Margin="0,0,5,0"/>
-        <TextBlock x:Name="textSlot3"
-                   Grid.Row="3" Grid.Column="1"
-                   Text="-" TextWrapping="Wrap"/>
-    </Grid>
-</UserControl>
