+++ Views/Controls/RelicData.xaml
@@ -1,0 +1,52 @@
+﻿<UserControl x:Class="NightreignRelicEditor.Views.Controls.RelicData"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             mc:Ignorable="d" d:DesignWidth="200">
+    <Border Padding="10">
+        <Grid HorizontalAlignment="Stretch">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="Auto"/>   <!-- CheckBox / Buttons -->
+                <ColumnDefinition Width="*"/>      <!-- Text -->
+            </Grid.ColumnDefinitions>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="Auto"/>     <!-- Relic header -->
+                <RowDefinition Height="Auto"/>     <!-- Slot 1 -->
+                <RowDefinition Height="Auto"/>     <!-- Slot 2 -->
+                <RowDefinition Height="Auto"/>     <!-- Slot 3 -->
+            </Grid.RowDefinitions>
+
+            <!-- Header -->
+            <CheckBox x:Name="checkRelic1Active"
+                      Grid.Row="0" Grid.ColumnSpan="2"
+                      Content="{Binding RelicName, RelativeSource={RelativeSource AncestorType=UserControl}}"
+                      IsChecked="True"
+                      FontWeight="Bold" />
+
+            <!-- Slot 1 -->
+            <Button x:Name="buttonClearSlot1"
+                    Grid.Row="1" Grid.Column="0"
+                    Content="x" Width="20" Margin="0,0,5,0"/>
+            <TextBlock x:Name="textSlot1"
+                       Grid.Row="1" Grid.Column="1"
+                       Text="-" TextWrapping="Wrap"/>
+
+            <!-- Slot 2 -->
+            <Button x:Name="buttonClearSlot2"
+                    Grid.Row="2" Grid.Column="0"
+                    Content="x" Width="20" Margin="0,0,5,0"/>
+            <TextBlock x:Name="textSlot2"
+                       Grid.Row="2" Grid.Column="1"
+                       Text="-" TextWrapping="Wrap"/>
+
+            <!-- Slot 3 -->
+            <Button x:Name="buttonClearSlot3"
+                    Grid.Row="3" Grid.Column="0"
+                    Content="x" Width="20" Margin="0,0,5,0"/>
+            <TextBlock x:Name="textSlot3"
+                       Grid.Row="3" Grid.Column="1"
+                       Text="-" TextWrapping="Wrap"/>
+        </Grid>
+    </Border>
+</UserControl>
